// playground-solid/src/App/PageComponents/PaintPicker/PaintPicker.types.ts
import type { SVGDefsColors, SVGDefsSamples } from "@thewaver/ss-components-solid";
import type { Size2d } from "@thewaver/ss-utils";

export type PaintKind = "none" | "pattern" | "timed" | "tracked";

export type PaintSampleKind = Exclude<PaintKind, "none">;

export type PaintSampleKey =
    | SVGDefsSamples.Pattern.SampleKey
    | SVGDefsSamples.Gradient.Timed.SampleKey
    | SVGDefsSamples.Gradient.Tracked.SampleKey;

export type Paint = {
    kind: PaintKind;
    key: PaintSampleKey;
    configDefs: Record<string, number | boolean>;
};

export type PaintSettings = {
    colors: SVGDefsColors;
    blurWidth?: number;
    animationDurationMs: number;
    iterationConfigKey: SVGDefsSamples.Iteration.SampleKey;
    cellSize: Size2d;
};

export type PaintSlot = {
    getPaint: () => Paint;
    getKind: () => PaintKind;
    setKind: (kind: PaintKind) => void;
    getKey: () => PaintSampleKey;
    setKey: (key: PaintSampleKey) => void;
    setConfigDef: (name: string, value: number | boolean) => void;
};

export type PagePaintPickerProps = {
    paintSlot: PaintSlot;
    name: string;
    label: string;
    hint: string;
};